% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{allBraidWords}
\alias{allBraidWords}
\title{Braid words of given length}
\usage{
allBraidWords(n, l)
}
\arguments{
\item{n}{number of strands, integer \code{>=2}}

\item{l}{length of the words}
}
\value{
A list of \code{braid} objects.
}
\description{
All braid words of the given length.
}
\examples{
allBraidWords(3, 2)
}
