% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{allPositiveBraidWords}
\alias{allPositiveBraidWords}
\title{Positive braid words of given length}
\usage{
allPositiveBraidWords(n, l)
}
\arguments{
\item{n}{number of strands, integer \code{>=2}}

\item{l}{length of the words}
}
\value{
A list of \code{braid} objects.
}
\description{
All positive braid words of the given length.
}
\examples{
allPositiveBraidWords(3, 4)
}
