% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{ArtinGenerators}
\alias{ArtinGenerators}
\alias{Sigma}
\alias{SigmaInv}
\title{Artin generators}
\usage{
Sigma(i)

SigmaInv(i)
}
\arguments{
\item{i}{index of the strand, a positive integer}
}
\value{
A vector of two integers.
}
\description{
A standard Artin generator of a braid: \code{Sigma(i)}
  represents twisting the neighbour strands \code{i} and \code{i+1},
  such that strand \code{i} goes \emph{under} strand \code{i+1}.
}
