% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidASCII.R
\name{braidASCII}
\alias{braidASCII}
\title{ASCII braid}
\usage{
braidASCII(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
No value is returned, just prints the ASCII figure.
}
\description{
Prints an ASCII figure of a braid.
}
\examples{
braid <- mkBraid(4, list(Sigma(1), SigmaInv(2)))
braidASCII(braid)
}
