% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{braidPermutation}
\alias{braidPermutation}
\title{Braid permutation}
\usage{
braidPermutation(braid)
}
\arguments{
\item{braid}{a \code{braid} object created with \code{\link{mkBraid}}}
}
\value{
A permutation.
}
\description{
Returns the left-to-right permutation associated to a braid.
}
\examples{
braid <- mkBraid(4, list(Sigma(2), SigmaInv(3), Sigma(3)))
braidPermutation(braid)
}
