% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{composeTwoBraids}
\alias{composeTwoBraids}
\title{Composition of two braids}
\usage{
composeTwoBraids(braid1, braid2)
}
\arguments{
\item{braid1, braid2}{\code{braid} objects with the same number of strands}
}
\value{
A \code{braid} object.
}
\description{
Composes two braids, doing free reduction on the result.
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
composeTwoBraids(braid, braid)
}
