% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{doubleSigma}
\alias{doubleSigma}
\title{Double generator}
\usage{
doubleSigma(n, s, t)
}
\arguments{
\item{n}{number of strands, integer \code{>=2}}

\item{s, t}{indices of two strands, \code{s < t}}
}
\value{
A \code{braid} object.
}
\description{
Generator \eqn{\sigma_{s,t}} in the Birman-Ko-Lee new
  presentation. It twists the strands \code{s} and \code{t} while going over
  all other strands (for \code{t=s+1}, this is \eqn{\sigma_s}).
}
\examples{
doubleSigma(5, 1, 3)
}
