% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{halfTwist}
\alias{halfTwist}
\title{Half-twist}
\usage{
halfTwist(n)
}
\arguments{
\item{n}{number of strands, integer \code{>=2}}
}
\value{
A \code{braid} object.
}
\description{
The (positive) half-twist of all the braid strands, usually
  denoted by \eqn{\Delta}.
}
\examples{
halfTwist(4)
}
