% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{isPermutationBraid}
\alias{isPermutationBraid}
\title{Whether a braid is a permutation braid}
\usage{
isPermutationBraid(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a braid is a permutation braid, that is,
  a positive braid where any two strands cross at most one, and positively.
}
\examples{
braid <- mkBraid(4, list(Sigma(2), SigmaInv(3), Sigma(3)))
isPermutationBraid(braid)
}
