% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{isPositiveBraidWord}
\alias{isPositiveBraidWord}
\title{Whether a braid is positive}
\usage{
isPositiveBraidWord(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a braid has only positive Artin generators.
}
\examples{
braid <- mkBraid(4, list(Sigma(2), SigmaInv(3), Sigma(3)))
isPositiveBraidWord(braid)
}
