% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{isPureBraid}
\alias{isPureBraid}
\title{Whether a braid is pure}
\usage{
isPureBraid(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a braid is pure, i.e. its permutation is trivial.
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
isPureBraid(braid)
}
