% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{mkBraid}
\alias{mkBraid}
\title{Make a braid}
\usage{
mkBraid(n, brgens)
}
\arguments{
\item{n}{number of strands, an integer, at least 2}

\item{brgens}{list of generators obtained with \code{\link{Sigma}} or
\code{\link{SigmaInv}}}
}
\value{
A \code{braid} object.
}
\description{
Make a braid.
}
\examples{
mkBraid(4, list(Sigma(2), SigmaInv(3)))
}
