% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{numberOfStrands}
\alias{numberOfStrands}
\title{Number of strands}
\usage{
numberOfStrands(braid)
}
\arguments{
\item{braid}{a \code{braid} object created with \code{\link{mkBraid}}}
}
\value{
An integer.
}
\description{
The number of strands of a braid.
}
