% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{permutationBraid}
\alias{permutationBraid}
\title{Permutation braid}
\usage{
permutationBraid(perm)
}
\arguments{
\item{perm}{a permutation}
}
\value{
A \code{braid} object.
}
\description{
Makes a permutation braid from a permutation.
}
\examples{
perm <- c(3, 1, 4, 2)
braid <- permutationBraid(perm)
isPermutationBraid(braid)
braidPermutation(braid)
}
