% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{strandLinking}
\alias{strandLinking}
\title{Linking number between two strands}
\usage{
strandLinking(braid, i, j)
}
\arguments{
\item{braid}{a \code{braid} object}

\item{i, j}{indices of two strands}
}
\value{
An integer.
}
\description{
The linking number between two strands of a braid.
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
strandLinking(braid, 1, 3)
}
\seealso{
\code{\link{linkingMatrix}} to get the linking numbers between
  all pairs of strands of the braid.
}
