% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{braidPermutation}
\alias{braidPermutation}
\title{Braid permutation}
\usage{
braidPermutation(braid)
}
\arguments{
\item{braid}{a \code{braid} object (e.g. created with \code{\link{mkBraid}})}
}
\value{
A permutation.
}
\description{
Returns the left-to-right permutation associated to a braid.
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
braidPermutation(braid)
}
