% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{freeReduceBraidWord}
\alias{freeReduceBraidWord}
\title{Free reduction of a braid}
\usage{
freeReduceBraidWord(braid)
}
\arguments{
\item{braid}{a \code{braid} object created with \code{\link{mkBraid}}}
}
\value{
A \code{braid} object.
}
\description{
Applies free reduction to a braid, i.e. removes pairs of
  consecutive generators inverse of each other.
}
\examples{
braid <- mkBraid(4, list(Sigma(2), SigmaInv(3), Sigma(3)))
freeReduceBraidWord(braid)
}
