% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{inverseBraid}
\alias{inverseBraid}
\title{Inverse braid}
\usage{
inverseBraid(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
A \code{braid} object.
}
\description{
The inverse of a braid (without performing reduction).
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
ibraid <- inverseBraid(braid)
composeTwoBraids(braid, ibraid)
}
