% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{linkingMatrix}
\alias{linkingMatrix}
\title{Linking matrix}
\usage{
linkingMatrix(braid)
}
\arguments{
\item{braid}{a \code{braid} object}
}
\value{
A matrix.
}
\description{
Linking numbers between all pairs of strands of a braid.
}
\examples{
braid <- mkBraid(4, c(2, -3, 3))
linkingMatrix(braid)
}
\seealso{
\code{\link{strandLinking}} to get the linking number between
  two strands of the braid.
}
