% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{mkBraid}
\alias{mkBraid}
\title{Make a braid}
\usage{
mkBraid(n, artingens)
}
\arguments{
\item{n}{number of strands, an integer, at least 2}

\item{artingens}{Artin generators given by a vector of non-zero
integers; a positive integer \code{i} corresponds to the
standard positive Artin generator of a braid which
represents twisting the neighbour strands \code{i} and \code{i+1},
such that strand \code{i} goes \emph{under} strand \code{i+1}; a
negative integer \code{-i} corresponds to the inverse.}
}
\value{
A \code{braid} object.
}
\description{
Make a braid.
}
\examples{
mkBraid(n = 4, c(2, -3))
}
